# docs/swagger.yaml
basePath: /
definitions:
...
  api.GetRoleByIdResponseDto:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updatedAt:
        type: string
    type: object
  api.GetTemplateResponseDto:
    properties:
      createdAt:
        type: string
      id:
        type: string
      text:
        type: string
      type:
        type: string
      updatedAt:
        type: string
    type: object
...
  api.ListTemplatesResponseDto:
    properties:
      id:
        type: string
      type:
        type: string
    type: object
  api.ListUsersInRoleResponseDto:
    properties:
      displayName:
        type: string
      id:
        type: string
      username:
        type: string
    type: object
...
  api.PagedTemplatesResponseDto:
    properties:
      items:
        items:
          $ref: '#/definitions/api.ListTemplatesResponseDto'
        type: array
      pagination:
        $ref: '#/definitions/api.Pagination'
    type: object
  api.PagedUsersInRoleResponseDto:
    properties:
      items:
        items:
          $ref: '#/definitions/api.ListUsersInRoleResponseDto'
        type: array
      pagination:
        $ref: '#/definitions/api.Pagination'
    type: object
...
  api.Pagination:
    properties:
      page:
        type: integer
      size:
        type: integer
      totalItems:
        type: integer
      totalPages:
        type: integer
    type: object
...
  api.PatchRoleRequestDto:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
...
info:
  contact: {}
  description: Open source OIDC/IDP server.
  title: Keyline API
paths:
...
  /api/virtual-servers/{virtualServerName}/projects/{projectSlug}/roles/{roleId}:
    delete:
      description: Delete a role by ID from a project
      parameters:
      - default: keyline
        description: Virtual server name
        in: path
        name: virtualServerName
        required: true
        type: string
      - description: Project slug
        in: path
        name: projectSlug
        required: true
        type: string
      - description: Role ID (UUID)
        in: path
        name: roleId
        required: true
        type: string
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
      security:
      - BearerAuth: []
      summary: Delete role
      tags:
      - Roles
    get:
      description: Get a role by its ID within a project.
      parameters:
      - default: keyline
        description: Virtual server name
        in: path
        name: virtualServerName
        required: true
        type: string
      - description: Project slug
        in: path
        name: projectSlug
        required: true
        type: string
      - description: Role ID (UUID)
        in: path
        name: roleId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GetRoleByIdResponseDto'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get role
      tags:
      - Roles
    patch:
      consumes:
      - application/json
      description: Update a role by ID within a project
      parameters:
      - default: keyline
        description: Virtual server name
        in: path
        name: virtualServerName
        required: true
        type: string
      - description: Project slug
        in: path
        name: projectSlug
        required: true
        type: string
      - description: Role ID (UUID)
        in: path
        name: roleId
        required: true
        type: string
      - description: Role data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.PatchRoleRequestDto'
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
        "404":
          description: Role not found
      security:
      - BearerAuth: []
      summary: Patch role
      tags:
      - Roles
...
  /api/virtual-servers/{virtualServerName}/templates:
    get:
      parameters:
      - default: keyline
        description: Virtual server name
        in: path
        name: virtualServerName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.PagedTemplatesResponseDto'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: List templates
      tags:
      - Templates
  /api/virtual-servers/{virtualServerName}/templates/{templateType}:
    get:
      parameters:
      - default: keyline
        description: Virtual server name
        in: path
        name: virtualServerName
        required: true
        type: string
      - description: Template type
        in: path
        name: templateType
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GetTemplateResponseDto'
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get template
      tags:
      - Templates
...
  /api/virtual-servers/{vsName}/projects/{projectSlug}/roles/{roleId}/users:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of users
      parameters:
      - default: keyline
        description: Virtual server name
        in: path
        name: vsName
        required: true
        type: string
      - description: Project slug
        in: path
        name: projectSlug
        required: true
        type: string
      - description: Role ID (UUID)
        in: path
        name: roleId
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: pageSize
        type: integer
      - description: Order by field
        in: query
        name: orderBy
        type: string
      - description: Order direction (asc|desc)
        in: query
        name: orderDir
        type: string
      - description: Search term
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.PagedUsersInRoleResponseDto'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: List users in role
      tags:
      - Roles
...
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"